refactor(aggregator): merge snapshotter test doubles together
to avoid clutturing the module with test tooling
